scripts: Scripts now on its own folder to ease including them in PATH
In order to access in an easy way to the py7slib scripts system-wide,
all the scripts have been placed in a sepparated folder that could
be added to system's PATH.

Import system of the scripts has been modified for allowing the
import of the py7slib modules. Also the Install script is modified
to generate an importable file containing the path of py7slib.
This script must be called before using any of the scripts the
first time.